Grade: Update comment   -Added Text box and "Upload Comment" button on "Teacher Gradebook: Student Details" page.   -Retrieved selected student comment from "Student" table.   -Displayed selected student comment on text box.   -Written a post method on clicks of "Upload Comment" button.   -Written a method on "Student" model to store data in database.   -Redirected on "Gradebook" page after successfully upload comment.  Teacher Gradebook: Manage Grade   -Removed exception on clicks of "Delete" button from "Manage Grade" page.   -Debug the "Manage Grade" page.   -Written a method on "Grade" model and deleted all grade.  Small Fixes   -Removed exception at the time of creating "Forum" if user doesn't select outcomes.   -Debug the "Add Forum" page.   -Written a condition at the time of storing outcome data in database.

+++ models/GbItems.php
@@ -68,8 +68,10 @@ class GbItems extends BaseImasGbitems
     }
     public static function deleteById($gradeId)
     {
-        $grade = GbItems::findOne($gradeId);
-        $grade->delete();
+        $grade = GbItems::find()->where(['id' => $gradeId])->one();
+        if($grade) {
+            $grade->delete();
+        }
     }
     public static function updateGrade($gradeId,$AssignValue,$temp)
     {
